src/requirements.txt
requests>=2.31.0
boto3>=1.34.0
botocore>=1.34.0
PyJWT>=2.8.0